pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask-restx = "^1.3.0"
peewee = "^3.17.6"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
psycopg2 = "^2.9.9"
flask-jwt-extended = "^4.6.0"
redis = "^5.1.1"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
